requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-import-export==1.2.0
gunicorn==19.9.0
lazy-object-proxy==1.4.2
Pillow==6.2.0
psycopg2==2.8.3
python-dotenv==0.10.3
whitenoise==4.1.4

